// client/src/app/tournament/tournament-entry/model/event-entry-status.enum.ts
export enum EventEntryStatus {
  NOT_ENTERED = 'NOT_ENTERED',
  ENTERED = 'ENTERED',
  ENTERED_WAITING_LIST = 'ENTERED_WAITING_LIST',
  PENDING_CONFIRMATION = 'PENDING_CONFIRMATION',
  PENDING_DELETION = 'PENDING_DELETION',
  PENDING_WAITING_LIST = 'PENDING_WAITING_LIST',
  RESERVED_WAITING_LIST = 'RESERVED_WAITING_LIST'
}