// client/src/app/tournament/tournament-entry/model/availability-status.enum.ts
/**
 * Availability status of each event
 */
export enum AvailabilityStatus {
  ALREADY_ENTERED = 'ALREADY_ENTERED',
  AVAILABLE_FOR_ENTRY = 'AVAILABLE_FOR_ENTRY',
  EVENT_FULL = 'EVENT_FULL',
  DISQUALIFIED_BY_AGE = 'DISQUALIFIED_BY_AGE',
  DISQUALIFIED_BY_GENDER = 'DISQUALIFIED_BY_GENDER',
  DISQUALIFIED_BY_RATING = 'DISQUALIFIED_BY_RATING',
  SCHEDULING_CONFLICT = 'SCHEDULING_CONFLICT',
  MAX_EVENTS_PER_DAY = 'MAX_EVENTS_PER_DAY',
  MAX_EVENTS_PER_TOURNAMENT = 'MAX_EVENTS_PER_TOURNAMENT'
}